need to use BigDecimal in java to map correctly to decimal values in mysql. updating options prices and package prices views and related files

+++ intellij_project/src/main/java/it/polimi/telco_webapp/entities/PackagePricesView.java
@@ -6,6 +6,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import org.eclipse.persistence.annotations.ReadOnly;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 
@@ -21,18 +22,18 @@ public class PackagePricesView {
     @Id
     private int package_id;
     @Column
-    private float period1_total;
+    private BigDecimal period1_total;
     @Column
-    private float period2_total;
+    private BigDecimal period2_total;
     @Column
-    private float period3_total;
+    private BigDecimal period3_total;
 
 
 
     public int getPackageId(){return package_id;}
-    public float getPeriod1Total(){return period1_total;}
-    public float getPeriod2Total(){return period2_total;}
-    public float getPeriod3Total() {return period3_total;}
+    public BigDecimal getPeriod1Total(){return period1_total;}
+    public BigDecimal getPeriod2Total(){return period2_total;}
+    public BigDecimal getPeriod3Total() {return period3_total;}
 
     /* Since this is a view, the data is read-only. No setters. */
 }
